+++ kroxylicious-kafka-message-tools/etc/module-checkstyle.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0"?>
+<!--
+
+    Copyright Kroxylicious Authors.
+
+    Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+
+-->
+<!DOCTYPE module PUBLIC
+        "-//Checkstyle//DTD Checkstyle Configuration 1.3//EN"
+        "https://checkstyle.org/dtds/configuration_1_3.dtd">
+
+<module name="Checker">
+    <module name="TreeWalker">
+        <module name="ImportControl">
+            <property name="file" value="${project.basedir}/etc/module-layering.xml"/>
+            <property name="path" value="src/main/java" />
+        </module>
+    </module>
+</module>
