+++ kroxylicious-kafka-message-tools/src/main/java/io/kroxylicious/kafka/transform/RecordMapper.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.filter.encryption.records;
+package io.kroxylicious.kafka.transform;
 
 import org.apache.kafka.common.record.Record;
 import org.apache.kafka.common.record.RecordBatch;
