+++ kroxylicious-kafka-message-tools/src/test/java/io/kroxylicious/kafka/transform/RecordStreamTest.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.filter.encryption.records;
+package io.kroxylicious.kafka.transform;
 
 import java.nio.ByteBuffer;
 import java.nio.CharBuffer;
